[Update] ChatExtension 추가
